requirements.txt
python-dotenv
pydantic

openai
anthropic
google-genai
xai-sdk
groq
ollama